requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
requests
